# openapi.json
{
  "components": {
    "schemas": {
      "AdoptionOut": {
        "description": "AI 권고 채택률. **사람이 판단한 것만 센다.**",
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          },
          "pending": {
            "title": "Pending",
            "type": "integer"
          },
          "rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate"
          },
          "reviewed": {
            "title": "Reviewed",
            "type": "integer"
          }
        },
        "required": [
          "reviewed",
          "accepted",
          "pending"
        ],
        "title": "AdoptionOut",
        "type": "object"
      },
      "AgentOut": {
        "properties": {
          "awaiting_confirmation": {
            "title": "Awaiting Confirmation",
            "type": "boolean"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalated": {
            "title": "Escalated",
            "type": "boolean"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "verified": {
            "default": false,
            "title": "Verified",
            "type": "boolean"
          }
        },
        "required": [
          "session_id",
          "response",
          "awaiting_confirmation",
          "escalated"
        ],
        "title": "AgentOut",
        "type": "object"
      },
      "ConfirmIn": {
        "properties": {
          "approved": {
            "default": true,
            "title": "Approved",
            "type": "boolean"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id"
        ],
        "title": "ConfirmIn",
        "type": "object"
      },
      "DecisionOut": {
        "description": "판단 결과. **`proceed` 에 따라 나머지 필드가 갈린다.**\n\n거절이면 `reason` 만, 진행이면 환불 정보가 실린다. 한 모델로 두되 서로\n배타적인 필드를 선택으로 남긴다 — 소비자가 `proceed` 를 보고 갈라야 한다는\n사실이 스키마에 드러나야 하기 때문이다. `dict` 로 두면 그 사실이 사라지고,\n화면은 환불 금액이 항상 있다고 가정하게 된다.",
        "properties": {
          "policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy"
          },
          "proceed": {
            "title": "Proceed",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "refund_amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refund Amount"
          },
          "refund_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refund Ratio"
          }
        },
        "required": [
          "proceed"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
      "DecisionRow": {
        "description": "원장 한 줄. **거절 사유가 필수다** — 이유 없는 거절은 콘솔에서 쓸모가 없다.",
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "agent": {
            "title": "Agent",
            "type": "string"
          },
          "cost": {
            "title": "Cost",
            "type": "integer"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "human_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Decision"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "unit",
          "agent",
          "action",
          "cost",
          "decision",
          "reason"
        ],
        "title": "DecisionRow",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessageIn": {
        "properties": {
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "session_id": {
            "minLength": 1,
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "message",
          "request_id"
        ],
        "title": "MessageIn",
        "type": "object"
      },
      "ReviewIn": {
        "properties": {
          "approve": {
            "title": "Approve",
            "type": "boolean"
          },
          "intervention_id": {
            "title": "Intervention Id",
            "type": "integer"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "reviewer": {
            "minLength": 1,
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "intervention_id",
          "reviewer",
          "approve"
        ],
        "title": "ReviewIn",
        "type": "object"
      },
      "ReviewOut": {
        "properties": {
          "blocked": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "executed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed"
          },
          "final_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Decision"
          },
          "human_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Decision"
          },
          "intervention_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intervention Id"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ReviewOut",
        "type": "object"
      },
      "SessionOut": {
        "description": "세션 현재 상태.\n\n예전에는 `dict` 를 그대로 돌려줬다. 그러면 OpenAPI 에 모양이 안 실리고,\n소비자는 응답 모양을 손으로 적을 수밖에 없다 — 그렇게 적힌 것은 서비스가\n바뀌어도 아무 데서도 안 걸린다.",
        "properties": {
          "awaiting_confirmation": {
            "title": "Awaiting Confirmation",
            "type": "boolean"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DecisionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "escalated": {
            "title": "Escalated",
            "type": "boolean"
          },
          "next_nodes": {
            "items": {
              "type": "string"
            },
            "title": "Next Nodes",
            "type": "array"
          },
          "response": {
            "title": "Response",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/TraceEntry"
            },
            "title": "Trace",
            "type": "array"
          }
        },
        "required": [
          "session_id",
          "awaiting_confirmation",
          "next_nodes",
          "response",
          "escalated",
          "trace"
        ],
        "title": "SessionOut",
        "type": "object"
      },
      "SummaryOut": {
        "properties": {
          "adoption": {
            "$ref": "#/components/schemas/AdoptionOut"
          },
          "by_agent": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "integer"
              },
              "type": "object"
            },
            "title": "By Agent",
            "type": "object"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "policy_pass_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Pass Rate"
          },
          "spent": {
            "title": "Spent",
            "type": "integer"
          }
        },
        "required": [
          "counts",
          "by_agent",
          "spent",
          "adoption"
        ],
        "title": "SummaryOut",
        "type": "object"
      },
      "TraceEntry": {
        "additionalProperties": true,
        "description": "트레이스 한 줄.\n\n**`node` 만 보장한다.** 나머지 키는 노드마다 다르고, 노드가 늘면 또 달라진다.\n전부 열거하면 노드를 하나 추가할 때마다 스키마가 따라 움직여야 하고, 그렇다고\n통째로 `dict` 로 두면 소비자가 `node` 조차 믿을 수 없다. 보장되는 것만\n보장하고 나머지는 열어 둔다.",
        "properties": {
          "node": {
            "title": "Node",
            "type": "string"
          }
        },
        "required": [
          "node"
        ],
        "title": "TraceEntry",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Agent Customer Support",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/orchestration/pending": {
      "get": {
        "description": "사람 승인 대기. 이미 판단한 건은 빠진다.",
        "operationId": "orchestration_pending_orchestration_pending_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionRow"
                  },
                  "title": "Response Orchestration Pending Orchestration Pending Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orchestration Pending"
      }
    },
    "/orchestration/recent": {
      "get": {
        "description": "최근 결정. **실행된 것만 보면 통제가 안 보인다.**",
        "operationId": "orchestration_recent_orchestration_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decision"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionRow"
                  },
                  "title": "Response Orchestration Recent Orchestration Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orchestration Recent"
      }
    },
    "/orchestration/review": {
      "post": {
        "description": "사람의 판단을 받는다. **승인해도 규칙은 그대로 걸린다.**\n\n대기 건은 묵는다. 그 사이 다른 에이전트가 그 단위를 가져갔다면 사람이\n승인했다는 이유로 한 단위 한 개입을 건너뛸 수 없다 — 그러면 사람의 손을 거친\n건이 오히려 측정을 깨는 통로가 된다. 그래도 **사람의 판단은 남긴다.**",
        "operationId": "orchestration_review_orchestration_review_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orchestration Review"
      }
    },
    "/orchestration/summary": {
      "get": {
        "description": "콘솔 첫 화면. **두 비율을 따로 낸다.**\n\n예산이 모자라 거절된 제안은 \"사람이 AI 를 안 믿었다\" 가 아니다. 하나로 묶으면\n예산을 줄이는 것만으로 채택률이 떨어지고, 그 숫자를 보고 모델을 의심하게 된다.",
        "operationId": "orchestration_summary_orchestration_summary_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Orchestration Summary"
      }
    },
    "/support/confirm": {
      "post": {
        "description": "고객 승인. **이 호출 없이는 상태 변경이 일어나지 않는다.**",
        "operationId": "confirm_support_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Confirm"
      }
    },
    "/support/messages": {
      "post": {
        "operationId": "send_message_support_messages_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Message"
      }
    },
    "/support/sessions/{session_id}": {
      "get": {
        "operationId": "get_session_support_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session"
      }
    }
  }
}